refactor region macro creation and add more behaviours
