+++ module/data/item/templates/regions/regionPropertiesData.js
@@ -1,8 +1,10 @@
+import regionBehaviours from './regionBehavioursData.js';
+
 const fields = foundry.data.fields;
 
 export default function regionProperties() {
     return {
         createRegionFromTemplate: new fields.BooleanField({ initial: false }),
-        applyMacroOnEnter: new fields.StringField({ initial: '' })
+        behaviours: new fields.SchemaField(regionBehaviours())
     };
 }
